changed seed.py updated browswer router fixed errors
